Rename user app to myuser and fix related field in other apps
