+++ BabaShop/BabaShop/settings.py
@@ -43,7 +43,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
 
     'blog',
-    'user',
+    'myuser',
     'order',
     'shop',
 ]
