+++ BabaShop/blog/templates/blog/category_posts.html
@@ -33,7 +33,7 @@
                                 {{ post.short_description }}
 
                                 <hr>
-                                Written by: {{ post.author }}
+                                Written by: {{ post.supplier }}
                                 <br>
                                 Created at: {{ post.created_at.date }}
