+++ BabaShop/blog/templates/blog/post_list.html
@@ -34,7 +34,7 @@ Posts
                     {{ post.short_description }}
 
                     <hr>
-                    Written by: {{ post.author }}
+                    Written by: {{ post.supplier }}
                     <br>
                     Created at: {{ post.created_at.date }}
